refactor(meta_package): simplifies macros
