+++ include/soccer-common/Utils/meta/meta.h
@@ -1,7 +1,7 @@
 #ifndef SOCCER_COMMON_META_H
 #define SOCCER_COMMON_META_H
 
-#include "soccer-common/Utils/meta/meta_package/meta_package.h"
+#include "property/property.h"
 #include "static_block/static_block.h"
 
 #endif // SOCCER_COMMON_META_H
